pyproject.toml
[tool.poetry]
name = "django_polls"
version = "0.1.0"
description = ""
authors = ["Pablo Moreno <pablomoreno.inf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
coverage = "^6.4.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
